% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{normalize_headers}
\alias{normalize_headers}
\title{Normalize the headers of a data frame in place}
\usage{
normalize_headers(df)
}
\arguments{
\item{df}{A data frame / data.table.}
}
\value{
\code{df} with normalized column names.
}
\description{
Normalize the headers of a data frame in place
}
