% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{print_scorecard}
\alias{print_scorecard}
\title{Pretty-print a scored trial to the console}
\usage{
print_scorecard(scores, trial_result)
}
\arguments{
\item{scores}{A \code{data.table} returned by \code{\link[=score_trial]{score_trial()}}.}

\item{trial_result}{The list returned by \code{\link[=run_trial]{run_trial()}}.}
}
\value{
\code{scores}, invisibly.
}
\description{
Pretty-print a scored trial to the console
}
