% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{TOOL_ACQUISITION}
\alias{TOOL_ACQUISITION}
\title{Default acquisition type (\code{"DIA"} / \code{"DDA"}) per tool}
\usage{
TOOL_ACQUISITION
}
\description{
Default acquisition type (\code{"DIA"} / \code{"DDA"}) per tool
}
