% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{TOOL_TRANSFORMS}
\alias{TOOL_TRANSFORMS}
\title{Whether to enable packed-column transform detection per tool}
\usage{
TOOL_TRANSFORMS
}
\description{
Whether to enable packed-column transform detection per tool
}
